requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
SpeechRecognition==3.10.0
pyttsx3==2.90
spacy==3.7.2
wolframalpha==5.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
gunicorn==21.2.0
pyaudio==0.2.14
assemblyai==0.44.3
